# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trading Bot API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: orders
    description: Order placement and history
  - name: logs
    description: Log entries
  - name: credentials
    description: API credential management
  - name: market
    description: Market data
  - name: analytics
    description: Advanced performance analytics
  - name: alerts
    description: Price alerts
  - name: journal
    description: Trade journal
  - name: strategies
    description: Trading strategy signals
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all placed orders
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: symbol
          in: query
          required: false
          schema:
            type: string
        - name: side
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place a new trading order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "200":
          description: Order result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a specific order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /logs:
    get:
      operationId: getLogs
      tags: [logs]
      summary: Get recent log entries from the trading bot log file
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: level
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"

  /credentials:
    get:
      operationId: getCredentialsStatus
      tags: [credentials]
      summary: Check if Binance API credentials are configured
      responses:
        "200":
          description: Credentials status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialsStatus"
    post:
      operationId: saveCredentials
      tags: [credentials]
      summary: Save Binance API credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsInput"
      responses:
        "200":
          description: Credentials saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialsStatus"

  /stats:
    get:
      operationId: getStats
      tags: [orders]
      summary: Get aggregated order statistics
      responses:
        "200":
          description: Order statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderStats"

  /market/prices:
    get:
      operationId: marketPrices
      tags: [market]
      summary: Live 24hr ticker for major pairs
      responses:
        "200":
          description: Price tickers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TickerData"

  /market/balance:
    get:
      operationId: accountBalance
      tags: [market]
      summary: Fetch account balance (requires credentials)
      responses:
        "200":
          description: Balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssetBalance"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /market/klines:
    get:
      operationId: getKlines
      tags: [market]
      summary: OHLCV candlestick data for a symbol
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: interval
          in: query
          required: false
          schema:
            type: string
            enum: [1m, 5m, 15m, 30m, 1h, 4h, 1d]
            default: "1h"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Candlestick data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Kline"

  /analytics:
    get:
      operationId: getAnalytics
      tags: [analytics]
      summary: Advanced performance analytics
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdvancedAnalytics"

  /analytics/daily-pnl:
    get:
      operationId: getDailyPnl
      tags: [analytics]
      summary: Daily P&L for calendar heatmap
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 90
      responses:
        "200":
          description: Daily P&L data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyPnl"

  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List all price alerts
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create a price alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
      responses:
        "200":
          description: Created alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"

  /alerts/{id}:
    delete:
      operationId: deleteAlert
      tags: [alerts]
      summary: Delete a price alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  /journal:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List trade journal entries
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Create a journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalInput"
      responses:
        "200":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"

  /journal/{id}:
    put:
      operationId: updateJournalEntry
      tags: [journal]
      summary: Update a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalInput"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
    delete:
      operationId: deleteJournalEntry
      tags: [journal]
      summary: Delete a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  /strategies/signals:
    get:
      operationId: getStrategySignals
      tags: [strategies]
      summary: Get real-time strategy signals for major pairs
      responses:
        "200":
          description: Strategy signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StrategySignal"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    TickerData:
      type: object
      required: [symbol, price, change, changePercent, high, low, volume]
      properties:
        symbol:
          type: string
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        high:
          type: number
        low:
          type: number
        volume:
          type: number

    AssetBalance:
      type: object
      required: [asset, balance, available]
      properties:
        asset:
          type: string
        balance:
          type: number
        available:
          type: number
        unrealizedPnl:
          type: number

    Kline:
      type: object
      required: [openTime, open, high, low, close, volume, closeTime]
      properties:
        openTime:
          type: number
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
        closeTime:
          type: number

    AdvancedAnalytics:
      type: object
      required: [winRate, lossRate, totalTrades, filledTrades, totalPnl, avgTrade, bestTrade, worstTrade, sharpeRatio, maxDrawdown, profitFactor, avgHoldingHours]
      properties:
        winRate:
          type: number
          description: Win rate as percentage (0-100)
        lossRate:
          type: number
        totalTrades:
          type: integer
        filledTrades:
          type: integer
        totalPnl:
          type: number
        avgTrade:
          type: number
        bestTrade:
          type: number
        worstTrade:
          type: number
        sharpeRatio:
          type: number
        maxDrawdown:
          type: number
          description: Max drawdown as percentage
        profitFactor:
          type: number
        avgHoldingHours:
          type: number

    DailyPnl:
      type: object
      required: [date, pnl, trades]
      properties:
        date:
          type: string
          description: ISO date string YYYY-MM-DD
        pnl:
          type: number
        trades:
          type: integer

    Alert:
      type: object
      required: [id, symbol, targetPrice, direction, triggered, createdAt]
      properties:
        id:
          type: integer
        symbol:
          type: string
        targetPrice:
          type: number
        direction:
          type: string
          enum: [ABOVE, BELOW]
        message:
          type: ["string", "null"]
        triggered:
          type: boolean
        createdAt:
          type: string

    AlertInput:
      type: object
      required: [symbol, targetPrice, direction]
      properties:
        symbol:
          type: string
        targetPrice:
          type: number
        direction:
          type: string
          enum: [ABOVE, BELOW]
        message:
          type: string

    JournalEntry:
      type: object
      required: [id, title, createdAt, updatedAt]
      properties:
        id:
          type: integer
        orderId:
          type: ["integer", "null"]
        title:
          type: string
        notes:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        sentiment:
          type: ["string", "null"]
          enum: [bullish, bearish, neutral, null]
        createdAt:
          type: string
        updatedAt:
          type: string

    JournalInput:
      type: object
      required: [title]
      properties:
        orderId:
          type: ["integer", "null"]
        title:
          type: string
        notes:
          type: string
        tags:
          type: string
        sentiment:
          type: string
          enum: [bullish, bearish, neutral]

    StrategySignal:
      type: object
      required: [symbol, signal, strategy, strength, reason, price, timestamp]
      properties:
        symbol:
          type: string
        signal:
          type: string
          enum: [BUY, SELL, NEUTRAL]
        strategy:
          type: string
        strength:
          type: number
          description: Signal strength 0-100
        reason:
          type: string
        price:
          type: number
        rsi:
          type: number
        sma20:
          type: number
        sma50:
          type: number
        timestamp:
          type: string

    OrderInput:
      type: object
      required: [symbol, side, orderType, quantity]
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [BUY, SELL]
        orderType:
          type: string
          enum: [MARKET, LIMIT, STOP_MARKET]
        quantity:
          type: number
        price:
          type: ["number", "null"]
        stopPrice:
          type: ["number", "null"]
        testMode:
          type: boolean

    Order:
      type: object
      required: [id, symbol, side, orderType, quantity, status, testMode, createdAt]
      properties:
        id:
          type: integer
        symbol:
          type: string
        side:
          type: string
        orderType:
          type: string
        quantity:
          type: number
        price:
          type: ["number", "null"]
        stopPrice:
          type: ["number", "null"]
        testMode:
          type: boolean
        status:
          type: string
        orderId:
          type: ["string", "null"]
        executedQty:
          type: ["string", "null"]
        avgPrice:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string

    OrderResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        testMode:
          type: boolean
        orderId:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        executedQty:
          type: ["string", "null"]
        avgPrice:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        order:
          $ref: "#/components/schemas/Order"

    LogEntry:
      type: object
      required: [timestamp, level, message]
      properties:
        timestamp:
          type: string
        level:
          type: string
        filename:
          type: ["string", "null"]
        message:
          type: string

    CredentialsStatus:
      type: object
      required: [configured, hasApiKey, hasApiSecret]
      properties:
        configured:
          type: boolean
        hasApiKey:
          type: boolean
        hasApiSecret:
          type: boolean
        message:
          type: string

    CredentialsInput:
      type: object
      required: [apiKey, apiSecret]
      properties:
        apiKey:
          type: string
        apiSecret:
          type: string

    OrderStats:
      type: object
      required: [total, successful, failed, byType, bySide, testOrders, realOrders]
      properties:
        total:
          type: integer
        successful:
          type: integer
        failed:
          type: integer
        testOrders:
          type: integer
        realOrders:
          type: integer
        byType:
          type: object
          additionalProperties:
            type: integer
        bySide:
          type: object
          additionalProperties:
            type: integer

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string