# lib/api-zod/src/generated/types/advancedAnalytics.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Trading Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface AdvancedAnalytics {
  /** Win rate as percentage (0-100) */
  winRate: number;
  lossRate: number;
  totalTrades: number;
  filledTrades: number;
  totalPnl: number;
  avgTrade: number;
  bestTrade: number;
  worstTrade: number;
  sharpeRatio: number;
  /** Max drawdown as percentage */
  maxDrawdown: number;
  profitFactor: number;
  avgHoldingHours: number;
}