# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertDirection } from './alertDirection';

export interface Alert {
  id: number;
  symbol: string;
  targetPrice: number;
  direction: AlertDirection;
  /** @nullable */
  message?: string | null;
  triggered: boolean;
  createdAt: string;
}