# lib/api-zod/src/generated/types/dailyPnl.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Trading Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyPnl {
  /** ISO date string YYYY-MM-DD */
  date: string;
  pnl: number;
  trades: number;
}