# lib/api-zod/src/generated/types/journalEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { JournalEntrySentiment } from './journalEntrySentiment';

export interface JournalEntry {
  id: number;
  /** @nullable */
  orderId?: number | null;
  title: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tags?: string | null;
  /** @nullable */
  sentiment?: JournalEntrySentiment;
  createdAt: string;
  updatedAt: string;
}